requirements2.txt
django
djangorestframework
dnspython
pandas
ipaddress
pytz
channels
channels_redis
paramiko
netmiko
psycopg2
django-filter
django-tenants
django_extensions
celery
radis
tenant-schemas-celery
drf-nested-routers
django-netfields
requests
django-polymorphic
ipaddress